// newsportal_backend/src/app/site_setting/site_setting.interface.ts
export interface ISiteSettingInterface {
  _id?: any;
  logo?: string;
  favicon?: string;
  title?: string;
  emergency_contact?: string;
  email?: string;
  facebook?: string;
  instagram?: string;
  you_tube?: string;
  watsapp?: string;
  twitter?: string;
  linkedin?: string;
  about_us?: string;
  privacy_policy?: string;
  terms_condition?: string;
  address?: string;
  advertisement?: string;
  cover_photo?: string;
}